embedding-service/requirements.txt
pika
requests
chromadb
python-dotenv
